Fix position to return floats instead of ints

+++ src/ticker/TickerItem.ts
@@ -38,8 +38,8 @@ export default class TickerItem extends Item {
         const output: [x: number, y: number] = [-9999, -9999];
 
         if (values) {
-            output[0] = parseInt(values[0]);
-            output[1] = parseInt(values[1]);
+            output[0] = parseFloat(values[0]);
+            output[1] = parseFloat(values[1]);
         }
 
         return output;
